Add get_salaries_by_languages() function
The function takes in a list of programming languages as input,
and returns a list of lists containing information about
the average salaries of job vacancies related to each language.
Each language list contains four pieces of information for that
language. The first element in this list is the language name,
the second element is the total number of vacancies found for
that language, the third element is the number of vacancies
with salary data available, and the fourth element is the mean
salary of all vacancies with available salary data.